Fixed headers.

+++ src/exhaustive.h
@@ -1,2 +1,2 @@
-int *wdp(int, int, int, int *, int *, int);
-int *wde(int, int, int, int, int *, int *, int);
+long *wdp(int, int, int, int *, int *, int);
+long *wde(int, int, int, int, int *, int *, int);
